+++ src/ports/inbound/video/youtube/downloader/__init__.py
@@ -3,7 +3,6 @@ from typing import Any, Callable
 
 from src.domain.entity.error.video import (
     DownloadingYouTubeVideoError,
-    GettingYouTubeVideoInfoError,
 )
 from src.domain.entity.video.youtube import DownloadedYouTubeVideo, YouTubeVideoInfo
 from src.domain.entity.video.download_status import (
@@ -45,19 +44,5 @@ class YouTubeVideoDownloaderInterface(ABC):
 
         Returns:
             DownloadingYouTubeVideoError | DownloadedYouTubeVideo: Either Error entity or a Success entity
-        """
-        raise Exception("This should be implemented from an adapter !")
-
-    @abstractmethod
-    def get_video_info_from_url(
-        self, *, url: str
-    ) -> GettingYouTubeVideoInfoError | YouTubeVideoInfo:
-        """Getting YouTube video info from a url
-
-        Args:
-            url (str): The url of the video we want to get info about
-
-        Returns:
-            GettingYouTubeVideoInfoError | YouTubeVideoInfo: Either Error entity or a Success entity
         """
         raise Exception("This should be implemented from an adapter !")
